+++ Event/PromotionEvent.h
@@ -7,9 +7,9 @@ class PromotionEvent : public Event
 {
 private:
 	int cargo_id;             // unique sequence number that identifies each cargo
-	int extra_money;         // extra cost for moving normal cargo to vip
+	float extra_money;         // extra cost for moving normal cargo to vip
 public:
-	PromotionEvent( MarsStation* pApp , int Eventtime_day , int Eventtime_hour , int id , int Extra_Money=0);
+	PromotionEvent( MarsStation* pApp , int Eventtime_day , int Eventtime_hour , int id , float Extra_Money=0);
 	virtual ~PromotionEvent(void);
