+++ Event/ReadyEvent.h
@@ -9,10 +9,10 @@ private:
 	double cargo_distance;    //The distance that the cargo will travel in kilometeres
 	int load_time;            //Time (in hours) to load or unload cargo
 	int cargo_id;             // unique sequence number that identifies each cargo
-	int cost;                 //cost of delievering the cargo
+	float cost;                 //cost of delievering the cargo
 	static int AutoP;        //represents number of days after which a normal cargo is Auto promoted to VIP cargo
 public:
-	ReadyEvent( MarsStation* pApp, int Eventtime_day , int Eventtime_hour , TYP type , double distance , int LoadTime , int id  , int Cost );
+	ReadyEvent( MarsStation* pApp, int Eventtime_day , int Eventtime_hour , TYP type , double distance , int LoadTime , int id  , float Cost );
 	virtual ~ReadyEvent(void);
 
 	TYP	get_cargo_type(char CargoType);
